+++ scripts/covid daily cases.R
@@ -13,35 +13,56 @@ pacman::p_load(
   ggplot2,
   dplyr,
   zoo,
-  lubridate
+  lubridate,
+  ggthemes,
+  ggtext
 )
 
 # IMPORT DATASET ----
 
 covid_cases_csv <- read.csv(url("https://api.coronavirus.data.gov.uk/v2/data?areaType=nation&areaCode=E92000001&metric=newCasesBySpecimenDate&format=csv"))
+
+# MUNGE DATA ----
+
+# set the date format
+covid_cases_csv$date = as.Date(covid_cases_csv$date, "%Y-%m-%d")
+
+# calculate seven day rolling average
 covid_cases_csv <- covid_cases_csv %>%
   dplyr::mutate(cases_07da = zoo::rollmean(newCasesBySpecimenDate, k = 7, fill = NA))
 
-# FORMAT DATA ----
+# restrict to last twelve months
+covid_cases_csv <- subset(covid_cases_csv, date>today() - months(12))
 
-covid_cases_csv$date = as.Date(covid_cases_csv$date, "%Y-%m-%d")
+# remove most recent five days
 less_recent_days <- Sys.Date() - 5
 less_seven_days <- less_recent_days - 7
+covid_cases_subset <- subset(covid_cases_csv, date < less_recent_days)
 
 # PLOT DATA ----
 
-covid_cases_plot <- ggplot(covid_cases_csv[which(covid_cases_csv$date>today() - months(12) & covid_cases_csv$date<less_recent_days),], aes(x = date, y = cases_07da)) + 
-  geom_point(shape = 1, colour = "red", size=2) +
-  geom_smooth(data=subset(covid_cases_csv, covid_cases_csv$date >= less_seven_days), method = "lm", colour = "black", size=0.5, fullrange=FALSE, se=FALSE) +
-  labs(caption = paste("Data from UK Health Security Agency / https://coronavirus.data.gov.uk. Plotted", Sys.time(), sep = " "))
-
-covid_cases_plot + 
+# set plot and geom
+covid_cases_plot <- ggplot() +
+  # plot cases
+  geom_point(data = covid_cases_subset, aes(x = date, y = cases_07da), shape = 1, colour = "red", size = 2) +
+  # plot trend line
+  geom_smooth(data = subset(covid_cases_csv, covid_cases_csv$date >= less_seven_days), aes(x = date, y = cases_07da), method = "lm", colour = "black", size=0.5, fullrange=FALSE, se=FALSE) +
+  # set x and y axis
   scale_y_continuous(trans = 'log10', breaks = c(1000,2000,5000,10000,20000,50000), position = "right") +
-  scale_x_date(date_labels = "%B %Y", date_breaks = "1 month") +
-  ggtitle("England covid cases - 7 day average by specimen date") +
+  scale_x_date(date_labels = "%b %y", date_breaks = "1 month") +
   xlab("Date") +
   ylab("Cases") +
-  theme_bw()
+  # set title, subtitle and caption
+  ggtitle("England covid cases - 7 day average by specimen date") +
+  labs(caption = paste0("Data from UK Health Security Agency / https://coronavirus.data.gov.uk. Plotted ", Sys.time(), sep = " "), subtitle = paste0("Daily numbers of new cases (people who have had at least one positive COVID-19 test result). Data are shown by the date the sample was taken from the person being tested.")) +
+  # set theme
+  theme_base() +
+  theme(
+    plot.subtitle = element_markdown(hjust = 0, vjust = 0, size = 12),
+    plot.caption = element_text(size = 11))
+
+# crete plot
+covid_cases_plot
 
 # save to daily file
 ggsave("output/daily_england_cases.png", width = 16.6, height = 8.65, units = "in")